+++ apps/dashboard/src/app/(dashboard)/(chain)/[chain_id]/[contractAddress]/(marketplace)/components/CancelDirectListing.tsx
@@ -1,5 +1,5 @@
-import { CancelTab } from "contract-ui/tabs/shared-components/cancel-tab";
 import type { ThirdwebContract } from "thirdweb";
+import { CancelTab } from "./cancel-tab";
 
 interface CancelDirectListingProps {
   contract: ThirdwebContract;
