+++ apps/dashboard/src/app/(dashboard)/(chain)/[chain_id]/[contractAddress]/(marketplace)/english-auctions/ContractEnglishAuctionsPage.tsx
@@ -2,30 +2,30 @@
 
 import type { ThirdwebContract } from "thirdweb";
 import { Heading } from "tw-components";
-import { CreateListingButton } from "../shared-components/list-button";
-import { DirectListingsTable } from "./components/table";
+import { CreateListingButton } from "../components/list-button";
+import { EnglishAuctionsTable } from "./components/table";
 
-interface ContractDirectListingsPageProps {
+interface ContractEnglishAuctionsProps {
   contract: ThirdwebContract;
 }
 
-export const ContractDirectListingsPage: React.FC<
-  ContractDirectListingsPageProps
+export const ContractEnglishAuctionsPage: React.FC<
+  ContractEnglishAuctionsProps
 > = ({ contract }) => {
   return (
     <div className="flex flex-col gap-6">
       <div className="flex flex-row items-center justify-between">
-        <Heading size="title.sm">Contract Listings</Heading>
+        <Heading size="title.sm">Contract Auctions</Heading>
         <div className="flex flex-row gap-4">
           <CreateListingButton
             contract={contract}
-            type="direct-listings"
-            createText="Create Direct Listing"
+            type="english-auctions"
+            createText="Create English Auction"
           />
         </div>
       </div>
 
-      <DirectListingsTable contract={contract} />
+      <EnglishAuctionsTable contract={contract} />
     </div>
   );
 };
